recaptcha

+++ Backend/api/routers/questions.js
@@ -3,6 +3,7 @@ const mongoose = require("mongoose");
 const jwt = require("jsonwebtoken");
 const bcrypt = require("bcrypt");
 const multer = require("multer");
+const request = require("request");
 const shortid = require("shortid");
 const cookieParser = require("cookie-parser");
 const nodemailer = require("nodemailer");
@@ -16,12 +17,25 @@ const checkAuth = require("../middleware/checkAuth");
 const checkAuthUser = require("../middleware/checkAuthUser");
 const checkAuthAdmin = require("../middleware/checkAuthAdmin");
 
-
 const router = express.Router();
 
 router.use(cookieParser());
 
 router.delete("/:questionId", async (req, res, next) => {
+	if (!req.body.captcha) {
+		return res.status(400).json({
+			message: "No recaptcha token",
+		});
+	}
+	const verifyURL = `https://www.google.com/recaptcha/api/siteverify?secret=${process.env.reCaptchaSecret}&response=${req.body.captcha}`;
+	request(verifyURL, (err, response, body) => {
+		body = JSON.parse(body);
+		if (!body.success || body.score < 0.4) {
+			return res.status(401).json({
+				message: "Something went wrong",
+			});
+		}
+	});
 	await Question.deleteOne({ _id: req.params.questionId })
 		.exec()
 		.then((result) => {
@@ -50,12 +64,24 @@ router.get("/all/:quizId", checkAuth, async (req, res, next) => {
 		});
 });
 
-router;
-
 router.post("/add", checkAuth, checkAuthAdmin, async (req, res, next) => {
 	await Quiz.findById(req.body.quizId)
 		.exec()
 		.then(async (result1) => {
+			if (!req.body.captcha) {
+				res.status(400).json({
+					message: "No recaptcha token",
+				});
+			}
+			const verifyURL = `https://www.google.com/recaptcha/api/siteverify?secret=${process.env.reCaptchaSecret}&response=${req.body.captcha}`;
+			request(verifyURL, (err, response, body) => {
+				body = JSON.parse(body);
+				if (!body.success || body.score < 0.4) {
+					res.status(401).json({
+						message: "Something went wrong",
+					});
+				}
+			});
 			new Question({
 				_id: new mongoose.Types.ObjectId(),
 				quizId: req.body.quizId,
@@ -87,12 +113,29 @@ router.patch(
 	checkAuth,
 	checkAuthAdmin,
 	async (req, res, next) => {
+		if (!req.body.captcha) {
+			res.status(400).json({
+				message: "No recaptcha token",
+			});
+		}
+		const verifyURL = `https://www.google.com/recaptcha/api/siteverify?secret=${process.env.reCaptchaSecret}&response=${req.body.captcha}`;
+		request(verifyURL, (err, response, body) => {
+			body = JSON.parse(body);
+			if (!body.success || body.score < 0.4) {
+				res.status(401).json({
+					message: "Something went wrong",
+				});
+			}
+		});
 		const updateOps = {};
 		var flag = 0;
 		for (const ops of req.body) {
 			updateOps[ops.propName] = ops.value;
 		}
-		await Question.updateOne({ _id: req.params.questionId }, { $set: updateOps })
+		await Question.updateOne(
+			{ _id: req.params.questionId },
+			{ $set: updateOps }
+		)
 			.exec()
 			.then((result) => {
 				res.status(200).json({
@@ -102,19 +145,35 @@ router.patch(
 	}
 );
 
-router.post('/csv',checkAuth,checkAuthAdmin,async(req,res,next)=>{
-  const {questions}= req.body
-  await Question.insertMany(questions).then((result)=>{
-    res.status(200).json({
-      message:"Success",
-      result
-    })
-  }).catch((err)=>{
-    res.status(400).json({
-      message:"Error",
-      error:err.toString()
-    })
-  })
-})
+router.post("/csv", checkAuth, checkAuthAdmin, async (req, res, next) => {
+	if (!req.body.captcha) {
+		return res.status(400).json({
+			message: "No recaptcha token",
+		});
+	}
+	const verifyURL = `https://www.google.com/recaptcha/api/siteverify?secret=${process.env.reCaptchaSecret}&response=${req.body.captcha}`;
+	request(verifyURL, (err, response, body) => {
+		body = JSON.parse(body);
+		if (!body.success || body.score < 0.4) {
+			return res.status(401).json({
+				message: "Something went wrong",
+			});
+		}
+	});
+	const { questions } = req.body;
+	await Question.insertMany(questions)
+		.then((result) => {
+			res.status(200).json({
+				message: "Success",
+				result,
+			});
+		})
+		.catch((err) => {
+			res.status(400).json({
+				message: "Error",
+				error: err.toString(),
+			});
+		});
+});
 
 module.exports = router;
